hoteles_function/requirements.txt
google-cloud-bigquery==3.17.1
google-cloud-pubsub==2.18.4
functions-framework==3.4.0
